* renamed package file to invoiceapp. * added the base layout class for templates.
