+++ app/src/main/AndroidManifest.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.fidisys.android.booksapp">
+          package="com.fidisys.android.invoiceapp">
 
     <application
-            android:name=".BooksApp"
+            android:name="com.fidisys.android.invoiceapp.InvoiceApp"
             android:allowBackup="true"
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name="com.fidisys.android.booksapp.view.activity.InvoiceEditActivity">
+        <activity android:name="com.fidisys.android.invoiceapp.view.activity.InvoiceEditActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
