--- app/src/main/java/com/fidisys/android/booksapp/repository/InvoiceRepository.kt
@@ -1,7 +1,0 @@
-package com.fidisys.android.booksapp.repository
-
-import com.fidisys.android.booksapp.repository.db.InvoiceDao
-
-class InvoiceRepository(private val dao: InvoiceDao) {
-
-}
