--- app/src/main/java/com/fidisys/android/booksapp/repository/api/InvoiceApi.kt
@@ -1,4 +1,0 @@
-package com.fidisys.android.booksapp.repository.api
-
-interface InvoiceApi {
-}
