--- app/src/main/java/com/fidisys/android/booksapp/view/fragment/InvoiceListFragment.kt
@@ -1,17 +1,0 @@
-package com.fidisys.android.booksapp.view.fragment
-
-import android.os.Bundle
-import android.view.View
-import androidx.fragment.app.Fragment
-
-class InvoiceListFragment : Fragment() {
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        setUpViews()
-    }
-
-    private fun setUpViews() {
-
-    }
-}
