+++ app/src/main/java/com/fidisys/android/invoiceapp/repository/InvoiceRepository.kt
@@ -1,0 +1,7 @@
+package com.fidisys.android.invoiceapp.repository
+
+import com.fidisys.android.invoiceapp.repository.db.InvoiceDao
+
+class InvoiceRepository(private val dao: InvoiceDao) {
+
+}
