+++ app/src/main/java/com/fidisys/android/invoiceapp/repository/api/InvoiceApi.kt
@@ -1,0 +1,4 @@
+package com.fidisys.android.invoiceapp.repository.api
+
+interface InvoiceApi {
+}
