+++ app/src/main/java/com/fidisys/android/invoiceapp/view/activity/InvoiceEditActivity.kt
@@ -1,0 +1,13 @@
+package com.fidisys.android.invoiceapp.view.activity
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.fidisys.android.invoiceapp.R
+
+class InvoiceEditActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_invoice_edit)
+    }
+}
