+++ app/src/main/java/com/fidisys/android/invoiceapp/view/fragment/InvoicePreviewFragment.kt
@@ -1,10 +1,10 @@
-package com.fidisys.android.booksapp.view.fragment
+package com.fidisys.android.invoiceapp.view.fragment
 
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
-import com.fidisys.android.booksapp.repository.data.Invoice
+import com.fidisys.android.invoiceapp.repository.data.Invoice
 import timber.log.Timber
 
 /**
