+++ app/src/main/res/layout/activity_invoice_edit.xml
@@ -13,12 +13,4 @@
         android:layout_height="match_parent"
         app:defaultNavHost="true"
         app:navGraph="@navigation/nav_graph" />
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:layout_width="wrap_content"
-            android:id="@+id/action_button_create_invoice"
-            app:srcCompat="@drawable/ic_add_black_24dp"
-            android:layout_margin="@dimen/activity_margin_medium"
-            android:layout_gravity="bottom|end"
-            android:layout_height="wrap_content"/>
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
