+++ app/src/main/res/layout/fragment_invoice_list.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              xmlns:app="http://schemas.android.com/apk/res-auto"
               android:orientation="vertical"
               android:layout_width="match_parent"
               android:layout_height="match_parent">
@@ -9,4 +10,12 @@
             android:id="@+id/invoice_recycler"
             android:layout_height="match_parent">
     </androidx.recyclerview.widget.RecyclerView>
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+            android:layout_width="wrap_content"
+            android:id="@+id/action_button_create_invoice"
+            app:srcCompat="@drawable/ic_add_black_24dp"
+            android:layout_margin="@dimen/activity_margin_medium"
+            android:layout_gravity="bottom|end"
+            android:layout_height="wrap_content"/>
 </LinearLayout>
