+++ app/src/main/res/layout/layout_logo_title_viw.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.appcompat.widget.LinearLayoutCompat
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+    <androidx.appcompat.widget.AppCompatTextView
+            android:text="Invoi"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"/>
+</androidx.appcompat.widget.LinearLayoutCompat>
